Using sling api for creating nodes
